fix: Add checks before uninstall

+++ k8s/uninstall-kind.sh
@@ -2,10 +2,16 @@
 
 set -euo pipefail
 
-kind delete cluster --name=kind
+if kind get clusters | grep -q '^kind$'; then
+  kind delete cluster --name=kind
+  echo "'kind' cluster deleted successfully"
+else
+  echo "'kind' cluster does not exist"
+fi
 
-echo "kind cluster deleted successfully"
-
-rm /usr/local/bin/kind
-
-echo "kind uninstalled successfully"
+if [ -e /usr/local/bin/kind ]; then
+  rm /usr/local/bin/kind
+  echo "kind uninstalled successfully"
+else
+  echo "/usr/local/bin/kind does not exist"
+fi
